requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
anthropic==0.18.1
PyQt6==6.6.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6
streamlit==1.32.0
requests==2.31.0 